+++ apps/admin/Controller/User/Unlock.php
@@ -51,6 +51,6 @@ class Unlock
      */
     private function id(): int
     {
-        return (int) Auth::getLogin()['id'];
+        return Auth::getLogin()['id'];
     }
 }
